requirements.txt
asgiref==3.3.1
Django==3.1.7
django-appconf==1.0.4
django-autoslug==1.9.8
django-el-pagination==3.3.0
django-endless-pagination==2.0
django-genericadmin==0.7.0
django-grappelli==2.14.3
django-imagekit==4.0.2
django-markdownx==3.0.1
djangorestframework==3.12.2
Markdown==3.3.3
pilkit==2.0
Pillow==8.1.0
psycopg2==2.8.6
pypandoc==1.5
pytz==2018.9
six==1.12.0
sqlparse==0.4.1
webencodings==0.5.1
